[refactor] :truck: rename common.ts to chatizalo.ts
